Update Kmeans_Examples.Rmd
Added an additional section on finding an "best" k using the "elbow method" and WSS. Also when we load in the Iris data I added a function to scale the features (x). This is good practice to ensure that the clusters are being formed on numbers in the same scale.